change timeout from uint8 to int

+++ lib/Embedded/SinexcelSer1000.cpp
@@ -393,7 +393,7 @@ int SinexcelSer1000::endPacket()
   }
 
   const int timeoutValue = 500;
-  uint8_t currenTimeout = 0;
+  int currenTimeout = 0;
 
   // wait for TX buffer to free
   while ((readRegister(REG_SR) & 0x04) != 0x04 && (currenTimeout < timeoutValue)) {
